Update flake.lock, and update zsh config for new style. kitty-save-session.sh update to dgrams.
Update the lockfile.
Zsh init now only uses mkOrder and setting an initContent instead of the
many different init* options.
Fix the golang overlay override to handle lambdas for args as well, and
move it to a separate file.
Update kitty-save-session to version with datagram support, and switch
to using that.