NixGL dGPU primary. Zed-editor added. Movable fleek clone dir.
Change the nixGL custom options to allow setting dGPU as primary when its the primary renderer, and
allow enabling vulkan support (both needed by Zed on systems where primary renderer is dGPU). Allow
configuring for systems that only have a dGPU too.

Add zed-editor config, which uses a mutable config from a subfolder by default, but can be configuredd
to use a static config generated by home-manager (Zed doesn't work with immutable config though).
Update the mutable config in the repo.
Use zed-editor with the nixGL wrapper, which works around a config bug in natively installed zed where
it fails to pick the primary renderer when there are multiple, which then forces a fallback to emulated
GPU.
Include option for when Vulkan support is insufficient, hiding the emulated GPU warning.

Add an option to all.nix that should be set per-host for the path to where the flake clone lives. Default
is the current required directory. This sets a FLEEK_CONFIG_DIR shell env variable, which is now used by
the fleek-* tools in bin/.  It also gets used to set the 'fleeks' alias, and when we need to out-of-store
symlink a folder from the cloned config to a known location (e.g. emacs and zed).